supervisor: Refactor into separate worker and master
sv struct was removed. Worker and Master are now separated to make the code
easier to reason about. supervisor now synchronously does networking setup
as the minion starts running, in preparation for a future patch that will
remove the need for the SupervisorInit flag.
